fix the disk usage report

+++ MyMonitorHub.Web/MyMonitorHub.Domain/Domain/Entity/Diskusage.cs
@@ -5,11 +5,11 @@ namespace MyMonitorHub.Domain.Entity
         public int DiskUsageId { get; set; }
         public int ItemId { get; set; }
         public System.DateTime Timestamp { get; set; }
-        public int DiskSize { get; set; }
-        public int CurUsedMB { get; set; }
-        public int MinUsedMB { get; set; }
-        public int MaxUsedMB { get; set; }
-        public int AvgUsedMB { get; set; }
+        public long DiskSize { get; set; }
+        public long CurUsedMB { get; set; }
+        public long MinUsedMB { get; set; }
+        public long MaxUsedMB { get; set; }
+        public long AvgUsedMB { get; set; }
         public virtual Item Item { get; set; } = null!;
     }
 }
